Fix UrlConversionTestCase Some attributes of a mocked object needed to be specified. It was causing a MagicMock to be used where a str or another type was expected. Also moved the test so that it can be specified by label, since the i18n label doesn't work as expected right now in the central server.
