+++ centralserver/i18n/tests/translation_tests.py
@@ -5,30 +5,12 @@ import os
 import polib
 import sys
 
-from mock import MagicMock
-
 from django.conf import settings
 from django.core.management import call_command
-from django.test import LiveServerTestCase, TestCase
+from django.test import LiveServerTestCase
 from django.utils import unittest
 
 from .. import POT_DIRPATH
-
-from centralserver.i18n.management.commands.update_language_packs import convert_aws_urls_to_localhost_urls
-
-class UrlConversionTestCase(TestCase):
-
-    def test_poentry_urls_converted(self):
-        """ poentry urls should be converted from aws urls to localhost urls """
-        poentry = MagicMock(autospec=polib.POEntry)
-        poentry.msgid = "I love https://something.aws.org/cat_picture.jpg"
-        poentry.msgstr = "Ich liebe https://something.aws.org/cat_picture.jpg"
-
-        poentry = convert_aws_urls_to_localhost_urls(poentry)
-
-        converted_url = "/content/khan/cat_picture.jpg"
-        self.assertIn(converted_url, poentry.msgid)
-        self.assertIn(converted_url, poentry.msgstr)
 
 class TranslationCommentTestCase(LiveServerTestCase):
     """
